+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All changes and updates, that are relevant for a user will be documented here
 
 ## Next Release
 
+### Features
+
+#### Transactions
+
+- Create transactions is now really only possible with the max amount of decimal places
+- Added some nice formatting for amount input in create transaction dialog
+
 ### Bug Fixes
 
 - Fixed: creating new profile without immediate save works now
 - Fixed: `Create Transaction` does not get the last used transaction type for creating transactions
 - Fixed: having same transaction multiple times in table view when saving after creating
+- Fixed: make it possible to have decimal numbers for transaction amounts
 
 <!-- insertion marker -->
 ## [0.0.4](https://github.com/repo/owner/releases/tag/0.0.4) - 2026-04-18